+++ src/Components/Home/aichatbot.jsx
@@ -1,0 +1,6 @@
+export default function App(){
+
+    return (
+        <h1>here goes the aichat bot</h1>
+    )
+}
