+++ src/Components/Home/newsletter.jsx
@@ -1,0 +1,6 @@
+export default function App(){
+
+    return (
+        <h1>here goes the newletter bot</h1>
+    )
+}
