+++ src/Components/Home/techstack.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+
+export default function App(){
+
+    return(
+        <div className="container mt-3">
+
+            <h1>Tech Stack</h1>
+            <div className="row justify-content-center">
+
+                <div className="col-lg-3">Html</div>
+                <div className="col-lg-3">CSS3</div>
+                <div className="col-lg-3">JavaScript</div>
+                <div className="col-lg-3">BootStrap</div>
+                <div className="col-lg-3">ReactJs</div>
+                <div className="col-lg-3">NodeJs</div>
+                <div className="col-lg-3">Sanity</div>
+                <div className="col-lg-3">Python</div>
+                <div className="col-lg-3">Netlify</div>
+            </div>
+        </div>
+    )
+}
